make changeset generation reusable
